Use ENV{ID_V4L_CAPABILITIES}==":capture:" to trigger on the correct port
UVC Cameras, such as the cameras from the Beta Teleop Kit, or popular
webcams, appear as two devices in the /dev/ folder. For example, the
gripper uvc camera from the teleop kit appears as /dev/video6 and
/dev/video7. The first is usually a capture device and can stream
imagery, and the second is usually purely for metadata and cannot
stream imagery. Therefore, it's very important that UVC camera UDEV
rules trigger on the first device. The critical attribute
differentiating these two devices is called ID_V4L_CAPABILITIES,
where video4linux seems to recognize which device can actually
stream imagery and marks it as "capture:". In this commit, we use
this attribute to trigger the UDEV rules on the correct device.